explainer: move explainer component to its own folder
