+++ .gitignore
@@ -1,4 +1,4 @@
 venv/
 .idea/
 __pycache__/
-.env
+explainer/.env
