+++ explainer/openai_client.py
@@ -1,0 +1,44 @@
+import openai
+
+
+class OpenAiClient:
+    MAX_TOKENS = 200
+    MODEL = "gpt-3.5-turbo"
+
+    def __init__(self, api_key):
+        """
+        Initializes an instance of the OpenAiClient class with the provided API key.
+        :param api_key: The API key for accessing the OpenAI service.
+        """
+        openai.api_key = api_key
+
+    @classmethod
+    def _set_prompt(cls, slide_txt: str) -> str:
+        """
+        returns a prompt to send to openai based on a given slide tex
+        :param slide_txt: the text of the slide
+        :return: the prompt
+        """
+
+        return f"""
+        Here is text extracted from a slide of a lecture presentation.
+        explain (and elaborate when needed) the content of each slide in a concise and coherent manner.
+        So that even a student who was not in the lecture will be able to understand the topic being discussed.
+
+        The text of the slide:
+        {slide_txt}
+        
+        Explanation of the topic following the text in the slide (6 - 7 sentences):
+        """
+
+    async def get_slide_explanation(self, slide_txt: str) -> str:
+        """
+        Retrieves the explanation for a slide from the OpenAI model.
+        :param slide_txt: The text of the slide.
+        :return: The explanation generated by the model.
+        """
+        response = await openai.ChatCompletion.acreate(
+            model=self.MODEL,
+            messages=[{"role": "user", "content": self._set_prompt(slide_txt)}],
+            max_tokens=self.MAX_TOKENS)
+        return response["choices"][0]["message"]["content"]
